Add dedicated de/activate_plugin_all methods
Add dedicated methods to bethesda.py for:
- activate_plugin_all
- deactivate_plugin_all

This allows us to handle the Union[int, str] for the index argument in
a way that allows static type checkers like `ty` to prove that we aren't
passing a string arg to functions that expect an int.

This is the same treatment as the last patch, but for bethesda.py.

Still need to do bethesda.py mod functions.